@@ -6,23 +6,58 @@ using System.Collections.ObjectModel;
 
 namespace BadMedicine.Dicom
 {
+    /// <summary>
+    /// Represents a whole DICOM Study (a collection of Series objects)
+    /// </summary>
     public class Study : IEnumerable<Series>
     {
+        /// <summary>
+        /// The Series objects within this Study
+        /// </summary>
         public IReadOnlyList<Series> Series{get;}
 
+        /// <summary>
+        /// The DicomDataGenerator which created this Study
+        /// </summary>
         public DicomDataGenerator Parent;
         
-        public DicomUID StudyUID{get; private set; }
+        /// <summary>
+        /// The DICOM UID of this Study
+        /// </summary>
+        public DicomUID StudyUID{get; }
+
+        /// <summary>
+        /// The timestamp on this Study
+        /// </summary>
         public DateTime StudyDate { get; internal set; }
 
-        public string StudyDescription {get; private set;}
+        /// <summary>
+        /// Free-text description of this Study
+        /// </summary>
+        public string StudyDescription {get; }
+        /// <summary>
+        /// The Accession Number for this Study, usually used to associate the study with clinical data in the RIS
+        /// </summary>
         public string AccessionNumber { get; }
+        /// <summary>
+        /// Starting time of the Study, empty if unknown
+        /// </summary>
         public TimeSpan StudyTime { get; }
 
+        /// <summary>
+        /// Count of Instances within this Study
+        /// </summary>
         public int NumberOfStudyRelatedInstances { get; }
 
-        private List<Series> _series = new List<Series>();
+        private readonly List<Series> _series = new List<Series>();
 
+        /// <summary>
+        /// Constructor for a new Study on a specified Person
+        /// </summary>
+        /// <param name="parent">The DicomDataGenerator creating this Study</param>
+        /// <param name="person">The Person representing this patient</param>
+        /// <param name="modalityStats">Statistical distributions to use</param>
+        /// <param name="r">Seeded PRNG to use</param>
         public Study(DicomDataGenerator parent, Person person, ModalityStats modalityStats, Random r)
         {
             /////////////////////// Generate all the Study Values ////////////////////
